2024/5/3-3

+++ StardewMUI 2023/核心/共享方法.vb
@@ -43,18 +43,18 @@ Public Class 共享方法
         End If
     End Function
 
-    Public Shared Function CompareVersion(Version1 As String, Version2 As String) As Integer
+    Public Shared Function CompareVersion(Version1 As String, Version2 As String) As Int128
         Dim cleanVersion1 As String = System.Text.RegularExpressions.Regex.Replace(Version1, "[^\d\.]", "")
         Dim cleanVersion2 As String = System.Text.RegularExpressions.Regex.Replace(Version2, "[^\d\.]", "")
         ' 将版本号拆分成数字数组
         Dim arrVersion1() As String = cleanVersion1.Split("."c)
         Dim arrVersion2() As String = cleanVersion2.Split("."c)
         ' 获取较长的版本号的长度
-        Dim maxLength As Integer = Math.Max(arrVersion1.Length, arrVersion2.Length)
+        Dim maxLength As Int128 = Math.Max(arrVersion1.Length, arrVersion2.Length)
         ' 逐位比较版本号
         For i As Integer = 0 To maxLength - 1
-            Dim num1 As Integer = If(i < arrVersion1.Length, Integer.Parse(arrVersion1(i)), 0)
-            Dim num2 As Integer = If(i < arrVersion2.Length, Integer.Parse(arrVersion2(i)), 0)
+            Dim num1 As Int128 = If(i < arrVersion1.Length, Int128.Parse(arrVersion1(i)), 0)
+            Dim num2 As Int128 = If(i < arrVersion2.Length, Int128.Parse(arrVersion2(i)), 0)
             If num1 < num2 Then
                 Return -1
             ElseIf num1 > num2 Then
